uniform nextWithoutConstraint()
